% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_BU13_panel.R
\docType{data}
\name{BU13_panel}
\alias{BU13_panel}
\title{Precomputed Bottom-up MST Assembly Example (BU13)}
\format{
An object of class \code{mstATA_panel}.
}
\source{
Generated using \code{mstATA::assembled_panels()} and \code{mstATA::solve_model()}with the HiGHS solver.
}
\usage{
BU13_panel
}
\description{
A precomputed multistage testing (MST) panel assembled using the
bottom-up strategy with the HiGHS solver. This object is included
to avoid long runtimes during vignette building and package examples.
}
\keyword{datasets}
