% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check_check_attribute_column.R
\name{check_attribute_column}
\alias{check_attribute_column}
\title{Check the item-level categorical and quantitative attribute in the itempool}
\usage{
check_attribute_column(itempool, attribute)
}
\arguments{
\item{itempool}{A \code{data.frame} containing the item pool. Must include metadata
for each item (e.g., item ID, difficulty, content, etc.).}

\item{attribute}{A string giving the column name for the item attribute in \code{itempool}.}
}
\value{
A vector of item attribute values with the same length as the item pool.
\itemize{
\item \strong{Quantitative attribute:} Returns a numeric vector.
Missing values (\code{NA}) are not allowed; a warning message will be issued if they are detected.
\item \strong{Categorical attribute:} Returns a factor vector indicating the category to which each item belongs.
Missing values (\code{NA}) are allowed and indicate that the corresponding item does not belong to any category for this attribute.
}
}
\description{
Confirm that the specified column exists in the item pool.
}
\keyword{internal}
