% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pivot_stimulus_map.R
\name{create_pivot_stimulus_map}
\alias{create_pivot_stimulus_map}
\title{Create Pivot–Stimulus Mapping}
\usage{
create_pivot_stimulus_map(
  itempool,
  item_id_col = "item_id",
  stimulus,
  pivot_item
)
}
\arguments{
\item{itempool}{A \code{data.frame} containing the item pool.}

\item{item_id_col}{A string giving the column name in \code{itempool}
that uniquely identifies items.}

\item{stimulus}{A string giving the column name in \code{itempool}
that identifies the stimulus or passage to which each item belongs.
Values may be character, or numeric (coerced to character).
Items not associated with a stimulus should be coded as \code{NA}.}

\item{pivot_item}{A string giving the column name in \code{itempool}.
Values must be character and non-missing values identify pivot items.
Each stimulus must have exactly one
pivot item. Empty strings or whitespace are treated as missing. See Details.}
}
\value{
A list with the following components:
\describe{

\item{\code{pivot_item_id}}{A character vector of pivot item IDs, one per
stimulus.}

\item{\code{stimulus_name}}{A character vector giving the stimulus
identifier associated with each pivot item. The order matches
\code{pivot_item_id}.}

\item{\code{stimulus_members}}{A named list. Each element contains the item
IDs belonging to a stimulus. Names match \code{stimulus_name}.}

\item{\code{numItems_stimulus}}{An integer vector giving the number of
items associated with each stimulus. This is \code{length(stimulus_members[[i]])}.}
}
}
\description{
Construct a stimulus–pivot item map from an item pool containing
stimulus identifiers and pivot-item indicators. This function extracts
the pivot item for each stimulus and identifies all items associated
with that stimulus.

This function is typically used as a \strong{preprocessing} step before applying
any constraints related with stimuli and stimulus-based items.
}
\details{
Formally, a pivot item is defined as an item that is selected
if and only if its corresponding stimulus is selected. In
practice, the pivot item is typically chosen as the item within a set
that best represents the stimulus—identified by content experts as having
the most representative content or desirable psychometric properties.
}
\examples{
itempool <- data.frame(
  item_id = 1:7,
  stimulus_id = c("S1","S1","S1","S2","S2", NA, NA),
  is_pivot   = c("P", NA, NA, "P", NA, NA, NA)
)

create_pivot_stimulus_map(
  itempool = itempool,
  item_id_col = "item_id",
  stimulus = "stimulus_id",
  pivot_item = "is_pivot"
)


}
