% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attribute_val.R
\name{get_attribute_val}
\alias{get_attribute_val}
\title{Get Item Categorical/Quantitative Attribute Value}
\usage{
get_attribute_val(itempool, attribute, cat_level = NULL)
}
\arguments{
\item{itempool}{A \code{data.frame} containing the item pool.}

\item{attribute}{A string giving the column name in \code{itempool} that represents
the categorical/quantitative attribute.}

\item{cat_level}{A optional character string for one category for the categorical attribute.
Default is NULL, indicating the attribute is a quantitative attribute.}
}
\value{
A vector of item attribute values. The same length as the item pool.
\itemize{
\item If the attribute is a quantitative attribute, return a numeric vector. NA is not allowed in the quantitative attribute.
\item If the attribute is a categorical attribute, return a binary integer vector with 1 indicating an item having the
categorical attribute, 0 indicating an item not having the categorical attribute. If an item has NA in the categorical attribute,
it will be coded as 0, indicating this item not having the categorical attribute.
}
}
\description{
Create a vector representing the attribute value.
}
\examples{
# Example 1, item categorical attribute
get_attribute_val(itempool = mini_itempool,attribute = "itemtype",cat_level = "MC")
# Example 2, item quantitative attribute
get_attribute_val(itempool = mini_itempool,attribute = "difficulty",cat_level = NULL)
# Example 3: stimulus categorical attribute
get_attribute_val(itempool = reading_itempool[!is.na(reading_itempool$pivot_item),],
                  attribute = "stimulus_type",cat_level = "history")
# Example 4: stimulus quantitative attribute
get_attribute_val(itempool = reading_itempool[!is.na(reading_itempool$pivot_item),],
                  attribute = "stimulus_words")
}
