% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_goal_programming_panel.R
\docType{data}
\name{goal_programming_panel}
\alias{goal_programming_panel}
\title{Precomputed MST Panel: Goal_Programming Formulation}
\format{
An object of class \code{mstATA_panel}.
}
\source{
Generated using \code{mstATA::assembled_panels()} and \code{mstATA::solve_model()}with the HiGHS solver.
}
\usage{
goal_programming_panel
}
\description{
A precomputed multistage testing (MST) panel assembled using the
goal_programming objective formulation (one_dev) described in the
\emph{Formulation for Multiple Objectives} vignette.
}
\details{
This object is included to avoid long solver runtimes during
vignette building and package checking.
}
\keyword{datasets}
