% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Hy13_panel.R
\docType{data}
\name{Hy13_panel}
\alias{Hy13_panel}
\title{Precomputed Hybrid MST Assembly Example}
\format{
An object of class \code{mstATA_panel}.
}
\source{
Generated using \code{mstATA::assembled_panels()} and \code{mstATA::solve_model()}with the HiGHS solver.
}
\usage{
Hy13_panel
}
\description{
A precomputed MST panel assembled using the hybrid strategy
with the HiGHS solver. Included to avoid long runtime during
vignette building.
}
\keyword{datasets}
