% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemcat_con.R
\name{itemcat_con}
\alias{itemcat_con}
\title{Generate Item-Level Constraints to Explicitly Select or Omit Specific Items}
\usage{
itemcat_con(
  x,
  item_ids,
  select = TRUE,
  which_module = NULL,
  which_pathway = NULL
)
}
\arguments{
\item{x}{An object of class \code{"mstATA_design"} created by \code{mst_design()}.}

\item{item_ids}{A numeric vector of item row indices, or
a character vector of item IDs from \code{x$ItemPool}.}

\item{select}{Logical. Indicates whether items in \code{item_ids} must be
\code{TRUE} (selected) or \code{FALSE} (not selected). Default = \code{TRUE}.}

\item{which_module}{Optional integer vector of module indices to which the constraints
apply.}

\item{which_pathway}{Optional integer vector of pathway indices to which the
constraints apply.}
}
\value{
An object of S3 class \code{"mstATA_constraint"} with named elements:
\describe{
\item{name}{A character vector indicating the specifications in each row of \code{A_binary}}
\item{specification}{A \code{data.frame} summarizing the constraint specification, including
the requirement name, attribute, constraint type, application level,
operator, and the number of constraint rows generated.}
\item{A_binary}{A sparse binary matrix representing the linear constraint coefficients.}
\item{A_real}{NULL for 'mstATA_constraint' object}
\item{operators}{A character vector of constraint operators, one per row of \code{A_binary}.}
\item{d}{A right-hand-side vector of 1s or 0s depending on whether \code{select} is \code{TRUE} or \code{FALSE}.}
\item{C_binary}{NULL for 'mstATA_constraint' object}
\item{C_real}{NULL for 'mstATA_constraint' object}
\item{sense}{NULL for 'mstATA_constraint' object}
}
}
\description{
This function generates linear constraints that require specific items to be
either must be \strong{selected} or \strong{not selected} in a multistage test (MST) panel
assembly.

If \code{item_module_eligibility} is supplied in \code{mst_design()} and the
requested item selection conflicts with module eligibility restrictions, an
error is thrown.
For details about \code{item_module_eligibility}, see \code{mst_design()}.

Constraints may be applied at the \code{"Module-level"}, \code{"Pathway-level"},
or \code{"Panel-level"}.

If \code{item_module_eligibility} is \code{NULL}, the number of constraints is:

\itemize{
\item \strong{Module-level:} when \code{which_module} is provided and
\code{which_pathway = NULL}

The number of constraints is
\strong{(number of items in \code{item_ids}) × (number of specified modules)}

\item \strong{Pathway-level:} when \code{which_pathway} is provided and
\code{which_module = NULL}

The number of constraints is
\strong{(number of items in \code{item_ids}) × (number of specified pathways)}

\item \strong{Panel-level:} both \code{which_module} and \code{which_pathway} are \code{NULL}

The number of constraints is
\strong{(number of items in \code{item_ids})}
}
}
\details{
\strong{1. Specification}

The constraint enforced by this function is:

\strong{Items listed in \code{item_ids} must (or must not) be selected.}

Key characteristics:
\itemize{
\item The attribute type is categorical (each item has a unique ID).
\item The attribute is defined at the item-level in the item pool.
\item Application levels:
\itemize{
\item \strong{Module-level}: enforce selection in specific module(s).
\item \strong{Pathway-level}: enforce selection in any module of a pathway.
\item \strong{Panel-level}: enforce selection somewhere in the panel.
}
}

\strong{2. Examples of Interpretation}

\itemize{

\item \strong{Item i must be selected in the routing module} (Module-level):
\code{item_ids = i}, \code{which_module = 1}

\item \strong{Item i must be selected in Stage 2 easy and medium modules}
(modules not belonging to the same pathway) (Module-level):
\code{which_module = c(2,3)}

\item \strong{Item i must be selected in the RM pathway} (Pathway-level):
\code{which_pathway = 2}

\item \strong{Item i must appear somewhere in the assembled panel}
(Panel-level):
\code{which_module = NULL}, \code{which_pathway = NULL}

}

\strong{3. Interaction with \code{panel_itemreuse_con()}}

When \code{overlap = FALSE} in \code{panel_itemreuse_con()}, each item can at most be selected once
in a panel.

Therefore:

\itemize{
\item \code{which_module} or \code{which_pathway} must be a scalar (not
a vector).
\item If both are \code{NULL}, the constraint simply enforces that the
item must appear somewhere in the panel, without specifying the
module/pathway.
}

When \code{overlap = TRUE} in \code{panel_itemreuse_con()}, each item can at most be selected once
in a pathway, but each item may be selected in multiple modules within the same stage.

Therefore:

\itemize{
\item \code{which_module} may be a vector (can not be modules that appear in the same pathway)
}
}
\section{Mathematical Formulation}{


Suppose the item pool contains (S - 1) stimulus-based item sets, indexed by
\eqn{s = 1, \ldots, S - 1}. Each stimulus has a designated pivot item,
indexed by \eqn{i_s^{*}}. In addition, the pool contains a set of discrete
(non–stimulus-based) items, which are represented by a dummy stimulus
\eqn{s = S} to allow a unified indexing scheme. Items belonging to stimulus
\eqn{s} are indexed as \eqn{i_s = 1, 2, \ldots, I_s}.

Suppose there are \eqn{M} modules in an MST panel. Let
\eqn{m = 1, \ldots, M} denote the module index.

\strong{1. Module-level item selection/not selection}

In specified module m: for all \eqn{i_s} specified in \code{item_ids}

\deqn{
  x_{i_s, m} = 1 \quad (\text{selected}), \qquad
  x_{i_s, m} = 0 \quad (\text{not selected}), \qquad
}

\strong{2. Pathway-level item selection/not selection}

In specified pathway r:  for all \eqn{i_s} specified in \code{item_ids}

\deqn{
  \sum_{m \in r} x_{i_s, m} = 1 \quad (\text{selected}), \qquad
  \sum_{m \in r} x_{i_s, m} = 0 \quad (\text{not selected}), \qquad
}

\strong{3. Panel-level item selection/not selection}

In a panel: for all \eqn{i_s} specified in \code{item_ids}

\deqn{
  \sum_{m} x_{i_s, m} = 1 \quad (\text{selected}), \qquad,
  \sum_{m} x_{i_s, m} = 0 \quad (\text{not selected}), \qquad
}

Here:
\itemize{
\item \eqn{x_{i_s, m}} is the binary decision variable indicating whether
item \eqn{i_s} is selected into module m.

\item \eqn{m \in r} denote modules belonging to pathway r.
}
}

\examples{
data("mini_itempool")
test_mstATA <- mst_design(
  itempool = mini_itempool,
  design = "1-3-3",
  module_length = c(4,2,2,2,2,2,2)
)

# Example 1: Item 2 must be selected in the routing module
itemcat_con(x = test_mstATA, item_ids = 2, select = TRUE, which_module = 1)

# Example 2: Item 2 must appear somewhere in the MST panel
itemcat_con(x = test_mstATA, item_ids = 2, select = TRUE)

# Example 3: Item 2 must be selected in Stage 2 easy and medium modules
itemcat_con(x = test_mstATA, item_ids = 2, select = TRUE, which_module = c(2,3))

# Example 4: Item 2 must be selected in the REE pathway
itemcat_con(x = test_mstATA, item_ids = 2, select = TRUE, which_pathway = 1)

# Example 5: Items 1 and 3 must be selected in the routing module
itemcat_con(x = test_mstATA, item_ids = c(1,3), select = TRUE, which_module = 1)

# Example 6: Items 1 and 3 must be selected in the REE pathway
itemcat_con(x = test_mstATA, item_ids = c(1,3), select = TRUE, which_pathway = 1)

# Example 7: Items 1 and 3 must appear somewhere in the panel
itemcat_con(x = test_mstATA, item_ids = c(1,3), select = TRUE)


# Example 8: With item_module_eligibility provided
mini_itempool <- mini_itempool[order(mini_itempool$difficulty), ]
item_module_eligibility <- list(
  `2` = 1:15,
  `3` = 11:25,
  `4` = 15:30
)

new_test_mstATA <- mst_design(
  itempool = mini_itempool,
  design = "1-3-3",
  module_length = c(6,4,4,4,3,3,3),
  item_module_eligibility = item_module_eligibility
)
itemcat_con(new_test_mstATA, item_ids = 2, select = TRUE, which_module = 2)
# error: item 2 is not eligible for the selection in module 3.
# itemcat_con(new_test_mstATA, item_ids = 2, select = TRUE, which_module = c(2,3))


}
