% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_panel.R
\docType{data}
\name{reading_panel}
\alias{reading_panel}
\title{Precomputed Stimulus-Based Assessment (MST Assembly Example)}
\format{
An object of class \code{mstATA_panel}.
}
\source{
Generated using \code{mstATA::assembled_panels()} and \code{mstATA::solve_model()}with the HiGHS solver.
}
\usage{
reading_panel
}
\description{
A precomputed MST panel assembled
with the HiGHS solver. Included to avoid long runtime during
vignette building.
}
\keyword{datasets}
