% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_TD12_panel.R
\docType{data}
\name{TD12_panel}
\alias{TD12_panel}
\title{Precomputed Top-Down MST Assembly Example (TD12)}
\format{
An object of class \code{mstATA_panel}.
}
\source{
Generated using \code{mstATA::assembled_panels()} and \code{mstATA::solve_model()}with the HiGHS solver.
}
\usage{
TD12_panel
}
\description{
A precomputed MST panel assembled using the top-down strategy
with the HiGHS solver. Included to avoid long runtime during
vignette building.
}
\keyword{datasets}
