% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_unary_minimax_panel.R
\docType{data}
\name{unary_minimax_panel}
\alias{unary_minimax_panel}
\title{Precomputed MST Panel: Unary_Minimax Formulation}
\format{
An object of class \code{mstATA_panel}.
}
\source{
Generated using \code{mstATA::assembled_panels()} and \code{mstATA::solve_model()}with the HiGHS solver.
}
\usage{
unary_minimax_panel
}
\description{
A precomputed multistage testing (MST) panel assembled using the
minimax objective formulation (one_dev) described in the
\emph{Formulation for Multiple Objectives} vignette.
}
\details{
This object is included to avoid long solver runtimes during
vignette building and package checking.
}
\keyword{datasets}
