% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helper_use_glpk.R
\name{use_glpk}
\alias{use_glpk}
\title{Solve model using GLPK}
\usage{
use_glpk(
  A,
  rhs,
  sense,
  obj,
  lb,
  ub,
  vtype,
  varname,
  modelsense,
  params,
  rowname
)
}
\description{
Uses GLPK to solve a linear or MILP model for test assembly
}
\keyword{internal}
