% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helper_use_gurobi.R
\name{use_gurobi}
\alias{use_gurobi}
\title{Solve model using gurobi}
\usage{
use_gurobi(
  A,
  rhs,
  sense,
  obj,
  lb,
  ub,
  vtype,
  varname,
  modelsense,
  params,
  rowname
)
}
\description{
Uses gurobi to solve a linear or MILP model for test assembly
}
\keyword{internal}
