% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helper_use_highs.R
\name{use_highs}
\alias{use_highs}
\title{Solve model using HiGHS}
\usage{
use_highs(
  A,
  rhs,
  sense,
  obj,
  lb,
  ub,
  vtype,
  varname,
  modelsense,
  params,
  rowname
)
}
\description{
Uses HiGHS to solve a linear or MILP model for test assembly
}
\keyword{internal}
