% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helper_use_lpsolve.R
\name{use_lpsolve}
\alias{use_lpsolve}
\title{Solve model using lpSolveAPI}
\usage{
use_lpsolve(
  A,
  rhs,
  sense,
  obj,
  lb,
  ub,
  vtype,
  varname,
  modelsense,
  params,
  rowname
)
}
\description{
Uses lpSolveAPI to solve a linear or MILP model for test assembly
}
\keyword{internal}
