% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helper_use_symphony.R
\name{use_symphony}
\alias{use_symphony}
\title{Solve model using symphony}
\usage{
use_symphony(
  A,
  rhs,
  sense,
  obj,
  lb = NULL,
  ub = NULL,
  vtype,
  varname,
  modelsense,
  params,
  rowname
)
}
\description{
Uses symphony to solve a linear or MILP model for test assembly
}
\keyword{internal}
