% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmst_pool.R
\docType{data}
\name{Rmst_pool}
\alias{Rmst_pool}
\title{Item pool from Rmst for examples}
\format{
A data frame with 440 rows and 14 columns:
\describe{
\item{Item_id}{Unique identifier for each item.}
\item{model}{IRT model specification for the item (e.g., \code{"3PL"},
\code{"GPCM"}, or \code{"GRM"}).}
\item{a}{Discrimination parameter (\eqn{a}) for the 3PL model.}
\item{b}{Difficulty parameter (\eqn{b}) for the 3PL model.}
\item{c}{Guessing parameter (\eqn{c}) for the 3PL model.}
\item{alpha}{Discrimination parameter \eqn{\alpha} for the
GPCM and GRM models.}
\item{delta1}{First step parameter \eqn{\delta_{1}} for the GPCM model.}
\item{delta2}{Second step parameter \eqn{\delta_{2}} for the GPCM model.}
\item{delta3}{Third step parameter \eqn{\delta_{3}} for the GPCM model.}
\item{beta1}{First threshold parameter \eqn{\beta_{1}} for the GRM model.}
\item{beta2}{Second threshold parameter \eqn{\beta_{2}} for the GRM model.}
\item{content}{Content category classification of the item.}
\item{time}{Item time.}
\item{group}{Item group identifier.}
}
}
\usage{
data(Rmst_pool)
}
\description{
A data frame containing 400 3PL items, 20 GPCM items, and 20 GRM items.
This item pool is adapted from the example
item pool used to demonstrate multistage test assembly in the
\pkg{Rmst} package.
}
\examples{
data(Rmst_pool)

}
\references{
Luo, X. (n.d.). \emph{Rmst: Multistage Testing Utilities in R}.
Retrieved from \url{https://github.com/xluo11/Rmst}
}
\keyword{datasets}
